preset_generator/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
numpy>=1.24.0
boto3>=1.34.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.24.0
python-multipart>=0.0.6
opencv-python>=4.8.0 